pyproject.toml
[project]
name = "whisper-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.4",
    "numpy>=2.3.5",
    "openai-whisper>=20250625",
    "torch>=2.9.1",
    "uvicorn>=0.38.0",
]
